+++ src/components/pages/apps/network/toolbars/home-toolbar.tsx
@@ -13,11 +13,15 @@ import { TEXT_PLOT, TEXT_SAVE_IMAGE } from '@/consts'
 import { useSVG } from '@/providers/svg-provider'
 
 import { useDialogs } from '@/components/dialogs/dialogs'
+import { DoubleNumericalInput } from '@/components/double-numerical-input'
+import { produce } from 'immer'
 import { NetworkDialog } from '../network-dialog'
+import { useNetworkSettings } from '../network-settings-store'
 
 export function HomeToolbar() {
   const { openDataFrames } = useOpenFiles({ mode: 'set' })
   const { saveAs } = useSVG()
+  const { settings, updateSettings } = useNetworkSettings()
 
   const { openCustom: openCustomDialog } = useDialogs()
 
@@ -60,6 +64,31 @@ export function HomeToolbar() {
           <PlayIcon variant="app-theme" />
           {TEXT_PLOT}
         </ToolbarColButton>
+      </ToolbarTabGroup>
+
+      <ToolbarTabGroup title="Plot Size">
+        <DoubleNumericalInput
+          h="sm"
+          v1={settings.plot.size.w}
+          placeholder="Width"
+          limit={[1, 5000]}
+          dp={0}
+          onNumChange1={(v) => {
+            updateSettings(
+              produce(settings, (draft) => {
+                draft.plot.size.w = v
+              })
+            )
+          }}
+          v2={settings.plot.size.h}
+          onNumChange2={(v) => {
+            updateSettings(
+              produce(settings, (draft) => {
+                draft.plot.size.h = v
+              })
+            )
+          }}
+        />
       </ToolbarTabGroup>
     </>
   )